Refact : MountainDetail Refactoring
